Object id deleted

+++ backend/building-security-system/src/main/java/com/building_security_system/models/Floor.java
@@ -4,7 +4,6 @@ import com.building_security_system.models.detectors.Detector;
 import java.util.List;
 
 public class Floor {
-    private int id;
+    private long id;
     private List<Detector> detectors;
-    //List<FloorFurniture> furniture;
 }
